requirements.txt
streamlit>=1.30,<2
python-docx>=1.1,<2
pypdf>=3.17,<4

numpy>=1.26,<3
pandas>=2.1,<3
scikit-learn>=1.4,<2
